extracted options dialog into generic resuable widget
